+++ src/main/java/com/greedy/mokkoji/api/club/controller/ClubController.java
@@ -2,13 +2,13 @@ package com.greedy.mokkoji.api.club.controller;
 
 import com.greedy.mokkoji.api.auth.controller.argumentResolver.AuthCredential;
 import com.greedy.mokkoji.api.auth.controller.argumentResolver.Authentication;
-import com.greedy.mokkoji.api.club.dto.club.request.ClubCreateRequest;
-import com.greedy.mokkoji.api.club.dto.club.request.ClubSearchCond;
-import com.greedy.mokkoji.api.club.dto.club.request.ClubUpdateRequest;
-import com.greedy.mokkoji.api.club.dto.club.response.ClubDetailResponse;
-import com.greedy.mokkoji.api.club.dto.club.response.ClubManageDetailResponse;
-import com.greedy.mokkoji.api.club.dto.club.response.ClubSearchResponse;
-import com.greedy.mokkoji.api.club.dto.club.response.ClubUpdateResponse;
+import com.greedy.mokkoji.api.club.dto.request.ClubCreateRequest;
+import com.greedy.mokkoji.api.club.dto.request.ClubSearchCond;
+import com.greedy.mokkoji.api.club.dto.request.ClubUpdateRequest;
+import com.greedy.mokkoji.api.club.dto.response.ClubDetailResponse;
+import com.greedy.mokkoji.api.club.dto.response.ClubManageDetailResponse;
+import com.greedy.mokkoji.api.club.dto.response.ClubsPaginationResponse;
+import com.greedy.mokkoji.api.club.dto.response.ClubUpdateResponse;
 import com.greedy.mokkoji.api.club.service.ClubService;
 import com.greedy.mokkoji.common.response.APISuccessResponse;
 import lombok.RequiredArgsConstructor;
@@ -34,7 +34,7 @@ public class ClubController {
     }
 
     @GetMapping
-    public ResponseEntity<APISuccessResponse<ClubSearchResponse>> getClubs(
+    public ResponseEntity<APISuccessResponse<ClubsPaginationResponse>> getClubs(
             @Authentication final AuthCredential authCredential,
             @ModelAttribute(value = "clubSearchCond") final ClubSearchCond clubSearchCond,
             @RequestParam(value = "page") final int page,
