+++ src/main/java/com/greedy/mokkoji/api/club/dto/request/ClubCreateRequest.java
@@ -1,15 +1,12 @@
-package com.greedy.mokkoji.api.club.dto.club.request;
+package com.greedy.mokkoji.api.club.dto.request;
 
 import com.greedy.mokkoji.enums.club.ClubAffiliation;
 import com.greedy.mokkoji.enums.club.ClubCategory;
 
-public record ClubUpdateRequest(
+public record ClubCreateRequest(
         String name,
         ClubCategory category,
         ClubAffiliation affiliation,
-        String description,
-        String clubMasterStudentId,
-        String logo,
-        String instagram
+        String clubMasterStudentId
 ) {
 }
