+++ src/main/java/com/greedy/mokkoji/api/club/dto/request/ClubSearchCond.java
@@ -1,4 +1,4 @@
-package com.greedy.mokkoji.api.club.dto.club.request;
+package com.greedy.mokkoji.api.club.dto.request;
 
 import com.greedy.mokkoji.enums.club.ClubAffiliation;
 import com.greedy.mokkoji.enums.club.ClubCategory;
