+++ src/main/java/com/greedy/mokkoji/api/club/dto/request/ClubUpdateRequest.java
@@ -1,12 +1,15 @@
-package com.greedy.mokkoji.api.club.dto.club.request;
+package com.greedy.mokkoji.api.club.dto.request;
 
 import com.greedy.mokkoji.enums.club.ClubAffiliation;
 import com.greedy.mokkoji.enums.club.ClubCategory;
 
-public record ClubCreateRequest(
+public record ClubUpdateRequest(
         String name,
         ClubCategory category,
         ClubAffiliation affiliation,
-        String clubMasterStudentId
+        String description,
+        String clubMasterStudentId,
+        String logo,
+        String instagram
 ) {
 }
