+++ src/main/java/com/greedy/mokkoji/api/club/dto/response/ClubDetailResponse.java
@@ -1,4 +1,4 @@
-package com.greedy.mokkoji.api.club.dto.club.response;
+package com.greedy.mokkoji.api.club.dto.response;
 
 import lombok.Builder;
 
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 @Builder
-public record ClubResponse(
+public record ClubDetailResponse(
         Long id,
         String name,
         String category,
@@ -15,9 +15,11 @@ public record ClubResponse(
         String recruitStartDate,
         String recruitEndDate,
         String logo,
-        Boolean isFavorite
+        Boolean isFavorite,
+        String instagram,
+        String recruitPost
 ) {
-    public static ClubResponse of(
+    public static ClubDetailResponse of(
             final Long id,
             final String name,
             final String category,
@@ -26,9 +28,11 @@ public record ClubResponse(
             final LocalDateTime recruitStartDate,
             final LocalDateTime recruitEndDate,
             final String logo,
-            final Boolean isFavorite) {
-
-        return ClubResponse.builder()
+            final Boolean isFavorite,
+            final String instagram,
+            final String recruitPost
+    ) {
+        return ClubDetailResponse.builder()
                 .id(id)
                 .name(name)
                 .category(category)
@@ -38,6 +42,8 @@ public record ClubResponse(
                 .recruitEndDate(recruitEndDate.format(DateTimeFormatter.ISO_LOCAL_DATE))
                 .logo(logo)
                 .isFavorite(isFavorite)
+                .instagram(instagram)
+                .recruitPost(recruitPost)
                 .build();
     }
 }
